backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
praw>=7.7.0
anthropic>=0.39.0
openai>=1.6.0
chromadb>=0.4.22
httpx>=0.25.0
pydantic>=2.5.0
